Configure VolumeMounts with subPath in ContainerBuilder
